chore: move and unify importer to fjs-viewer
