+++ packages/form-js-viewer/src/core/index.js
@@ -1,16 +1,19 @@
 import EventBus from './EventBus';
 import Validator from './Validator';
+import Importer from './Importer';
+import FieldFactory from './FieldFactory';
 import FormFieldRegistry from './FormFieldRegistry';
 import FormLayouter from './FormLayouter';
 
-import importModule from '../import';
 import renderModule from '../render';
 
-export { FormFieldRegistry, FormLayouter };
+export { Importer, FieldFactory, FormFieldRegistry, FormLayouter };
 
 export default {
-  __depends__: [ importModule, renderModule ],
+  __depends__: [ renderModule ],
   eventBus: [ 'type', EventBus ],
+  importer: [ 'type', Importer ],
+  fieldFactory: [ 'type', FieldFactory ],
   formFieldRegistry: [ 'type', FormFieldRegistry ],
   formLayouter: [ 'type', FormLayouter ],
   validator: [ 'type', Validator ]
