--- packages/form-js-viewer/src/import/index.js
@@ -1,5 +1,0 @@
-import Importer from './Importer';
-
-export default {
-  importer: [ 'type', Importer ]
-};
