+++ packages/form-js-viewer/src/index.js
@@ -1,6 +1,6 @@
 import Form from './Form';
 
-export { FormFieldRegistry, FormLayouter } from './core';
+export { FormFieldRegistry, FormLayouter, Importer, FieldFactory } from './core';
 export * from './render';
 export * from './util';
 export * from './features';
